Added project read_nop_test
Added a test program to check the possibility of
masking the delay of a read through the RoCC
interface with the insertion of independent
instructions and a NOP used to allow the writeback
of the data coming from the coprocessor.

The working configuration is:
	ROCC_READ a5, reg	//  a5 <- RoCC[reg]
	IndepInstr1
	IndepInstr2
	NOP
	IndepInstr3
	IndepInstr4
	IndepInstr5
	ADD a4, a4, a5		// use the value

where IndepInstrX are instruction that have no
data dependency with the ROCC_READ instruction.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   Makefile.def
#	new file:   accum_test/read_nop_test/Makefile
#	new file:   accum_test/read_nop_test/main.c
#